Rename CompositeJsonFormatter into AbstractCompositeJsonFormatter
Rename `CompositeJsonFormatter` into `AbstractCompositeJsonFormatter` to comply with the "abstract" naming convention.
