+++ src/main/java/net/logstash/logback/layout/LoggingEventCompositeJsonLayout.java
@@ -15,7 +15,7 @@
  */
 package net.logstash.logback.layout;
 
-import net.logstash.logback.composite.CompositeJsonFormatter;
+import net.logstash.logback.composite.AbstractCompositeJsonFormatter;
 import net.logstash.logback.composite.JsonProviders;
 import net.logstash.logback.composite.loggingevent.LoggingEventCompositeJsonFormatter;
 import net.logstash.logback.composite.loggingevent.LoggingEventJsonProviders;
@@ -26,7 +26,7 @@ import ch.qos.logback.core.joran.spi.DefaultClass;
 public class LoggingEventCompositeJsonLayout extends CompositeJsonLayout<ILoggingEvent> {
     
     @Override
-    protected CompositeJsonFormatter<ILoggingEvent> createFormatter() {
+    protected AbstractCompositeJsonFormatter<ILoggingEvent> createFormatter() {
         return new LoggingEventCompositeJsonFormatter(this);
     }
